What a mess; most lin. alg in here is prob wrong.
Backface culling looks like it worked... sort of.  And then trying to move the camera around, weird unexpected things started to happen (rotation of the world, when all I did was move the camera's eye origin... nope!).

Time to put this aside, brush the cobwebs off the mathematics part of my brain with some studying, and come back at a (much?) later time.