fix M4*V3 for non-1.0 W values
this seems to be what every other library does for M4*V3 operations.  I still feel like I'm missing something as to *why*, the deeper reasoning aside from the very general "multiplying a 4X4 Matrix with a 3-element Vector is certain to muddle up things because it's not as 'natural' as M3*V3 or M4*V4" so this checks the value of a calculated 4th element and adjusts xyz accordingly.

This did fix a couple inconsistencies elsewhere.  I wonder what else I'm still overlooking