split triangle draw function with a 2x-called draw-subtriangle function
Simplifies and makes it more readable and maintainable.

The two loops in original function were more or less the same, with only a couple incidental values different.  It's now a bit more readable, and was able to snip off a couple of variables in the primary triangle draw function as a result.